squash! WIP - Isolate blob URLs https://bugs.webkit.org/show_bug.cgi?id=241318
XXX SECURITY

Any frame may access a blob using a public URL reference, regardless of where
that blob was created. This provides a cross-site tracking vector. This
situation is complicated by the fact that blobs can be retrieved through
multiple paths. In this patch we restrict access to a blob by ensuring that the
blob is created and accessed only within the same top-origin context.