Strip StoragePartition from cookies returned by WKHTTPCookieStore API https://bugs.webkit.org/show_bug.cgi?id=292351 rdar://147233516
Reviewed by NOBODY (OOPS!).

The WKHTTPCookieStore API returns a NSHTTPCookie. Beginning in iOS 18.4, WebKit
enabled support for CHIPS, e.g., opt-in partitioned cookies. This feature is
intended to support partitioned third-party cookies, but it also supports
partitioned first party cookies. This is an additional compatibility affordance
for sites that may exist in both first and third party contexts. As such, when
a first party cookie is "partitioned", its NSHTTPCookie will contain the
"StoragePartition" property. This is internal state that does not need to be
exposed to application that use the WKHTTPCookieStore API, and it can cause
breakage in some situations.

This change removes the cookie property before the cookie is returned to the
application.

* Source/WebCore/platform/network/cocoa/NetworkStorageSessionCocoa.mm:
(WebCore::removeCookiePartition):
(WebCore::nsCookiesToCookieVector):
(WebCore::NetworkStorageSession::getAllCookies):
(WebCore::NetworkStorageSession::getCookies):
(WebCore::NetworkStorageSession::domCookiesForHost):
(WebCore::NetworkStorageSession::registerCookieChangeListenersIfNecessary):
* Source/WebKit/NetworkProcess/Cookies/WebCookieManager.cpp:
(WebKit::WebCookieManager::getCookies):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKHTTPCookieStore.mm:
(TEST(WKHTTPCookieStore, PartitionedCookieShouldNotHavePartitionProperty)):